refactor: reorganize XGo executor modules
