+++ spx-gui/build-wasm.sh
@@ -17,5 +17,5 @@ cp ../tools/ispx/ispx.wasm src/assets/wasm/ispx.wasm
 
 # Build the default executor bundle. It composes xgoexec with the Tutorial
 # framework; other framework bundles can be added without changing xgoexec.
-( cd ../tools/xgoexec-with-tutorial && ./build.sh )
-cp ../tools/xgoexec-with-tutorial/xgoexec.wasm src/assets/wasm/xgoexec.wasm
+( cd ../tools/xgoexec-bundle && ./build.sh )
+cp ../tools/xgoexec-bundle/xgoexec.wasm src/assets/wasm/xgoexec.wasm
