+++ tools/xgoexec-bundle/main.go
@@ -7,12 +7,12 @@ import (
 	"syscall/js"
 
 	tutorialbinding "github.com/goplus/builder/tools/tutorial/binding"
-	"github.com/goplus/builder/tools/xgoexec/core"
+	"github.com/goplus/builder/tools/xgoexec"
 )
 
-var runtime = core.NewRuntime(
-	map[string]core.Framework{tutorialbinding.Name: tutorialbinding.Binding{}},
-	core.RuntimeHooks{
+var runtime = xgoexec.NewRuntime(
+	map[string]xgoexec.Framework{tutorialbinding.Name: tutorialbinding.Binding{}},
+	xgoexec.RuntimeHooks{
 		Error: func(phase, message string) { js.Global().Call("xbuilder_xgoexec_error", phase, message) },
 		Exit:  func(reason string) { js.Global().Call("xbuilder_xgoexec_exit", reason) },
 	},
@@ -63,12 +63,12 @@ func dispatchEvent(this js.Value, args []js.Value) any {
 		if len(args) < 2 {
 			return fmt.Errorf("missing event name or payload")
 		}
-		return core.DispatchEvent(args[0].String(), []byte(args[1].String()))
+		return xgoexec.DispatchEvent(args[0].String(), []byte(args[1].String()))
 	})
 }
 func resolveCapabilityCall(this js.Value, args []js.Value) any {
 	if len(args) >= 3 {
-		core.ResolveCapabilityCall(uint64(args[0].Int()), args[1].String(), args[2].String())
+		xgoexec.ResolveCapabilityCall(uint64(args[0].Int()), args[1].String(), args[2].String())
 	}
 	return nil
 }
