+++ tools/xgoexec/framework.go
@@ -1,4 +1,4 @@
-package core
+package xgoexec
 
 // Framework supplies the build-time class-framework registration for a caller
 // selected XGo framework. Framework implementations deliberately live outside
